requirements.txt
streamlit
opencv-python
pillow
numpy
openai
python-dotenv
